host-bridgeの間違った接続の修正
